+++ research-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # rc: minor bump
 
+- Update the research app to allow selecting sources in either sky or planet-like
+  (Earth or Planet) modes (#369, @Carifio24).
+- Update the research app to be compatible with engine updates to screen <--> world
+  conversions (#378, @Carifio24).
+- Bump webpack from 5.94.0 to 5.104.1 (#351)
+
+
+# @wwtelescope/research-app 0.18.0 (2025-07-18)
+
 - Fix a bug where the research app's embedded Finder Scope could fail to draw its
   crosshairs (#333, @Carifio24). This could only happen after the Finder Scope was enabled,
   disabled, and re-enabled via the research app settings.